词法分析

+++ src/main/java/miniplc0java/instruction/Instruction.java
@@ -4,14 +4,14 @@ import java.util.Objects;
 
 public class Instruction {
     private Operation opt;
-    Integer x;
+    long x;
 
     public Instruction(Operation opt) {
         this.opt = opt;
         this.x = 0;
     }
 
-    public Instruction(Operation opt, Integer x) {
+    public Instruction(Operation opt, long x) {
         this.opt = opt;
         this.x = x;
     }
@@ -44,11 +44,11 @@ public class Instruction {
         this.opt = opt;
     }
 
-    public Integer getX() {
+    public long getX() {
         return x;
     }
 
-    public void setX(Integer x) {
+    public void setX(long x) {
         this.x = x;
     }
